should fix types to use registered routes
