+++ src/types/index.ts
@@ -10,4 +10,5 @@ export * from './routeWithParams'
 export * from './router'
 export * from './routerRoute'
 export * from './routes'
+export * from './routesMap'
 export * from './utilities'
