+++ src/types/register.ts
@@ -1,5 +1,7 @@
 import { Router } from '@/types/router'
 import { RouterRoutes } from '@/types/routerRoute'
+import { RoutesMap } from '@/types/routesMap'
+import { RouteWithParams } from '@/types/routeWithParams'
 
 export interface Register {
   // router: Router
@@ -22,3 +24,6 @@ export type RegisteredRejectionType = Register extends { rejections: infer TReje
 export type RegisteredRouterState = Register extends { state: infer TState }
   ? TState
   : {}
+
+export type RegisteredRouteMap = RoutesMap
+export type RegisteredRouteWithParams<T extends keyof RegisteredRouteMap> = RouteWithParams<RegisteredRoutes, T>
